+++ packages/generator-suit/CHANGELOG.md
@@ -1,5 +1,6 @@
 === HEAD
 
+* Add 'suit-test' to generated package's dev dependencies.
 * Switch to using Component(1) for development.
 
 === 0.1.1 (October 19, 2013)